Extract MainWindowViewModel from MainWindow code-behind
