+++ SentryReplay/MainWindow.xaml
@@ -8,7 +8,7 @@
         Title="Sentry Replay"
         MinWidth="1200"
         MinHeight="700"
-        d:DataContext="{d:DesignInstance Type=local:MainWindow, IsDesignTimeCreatable=False}"
+        d:DataContext="{d:DesignInstance Type=local:MainWindowViewModel, IsDesignTimeCreatable=False}"
         Background="{DynamicResource SystemFillColorSolidNeutralBackgroundBrush}"
         Closing="Window_Closing"
         ContentRendered="Window_ContentRendered"
