+++ SentryReplay/MainWindow.xaml.cs
@@ -1,252 +1,41 @@
 using System.ComponentModel;
 using System.Diagnostics;
-using System.IO;
-using System.Reflection;
-using System.Runtime.InteropServices;
 using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Input;
 using System.Windows.Navigation;
-using System.Windows.Threading;
-using CommunityToolkit.Mvvm.ComponentModel;
-using CommunityToolkit.Mvvm.Input;
 using FlyleafLib.Controls.WPF;
-using Microsoft.Win32;
 using Serilog;
 
 namespace SentryReplay;
 
 /// <summary>
-/// Main WPF window, player state, and Flyleaf host layout.
+/// Main WPF window. Owns only view concerns: window lifecycle, Flyleaf host layout, the seek
+/// slider input plumbing, and search-box focus. All state, commands, and orchestration live in
+/// <see cref="MainWindowViewModel"/>.
 /// </summary>
-[INotifyPropertyChanged]
 public partial class MainWindow : Window
 {
-    private const string GridCameraView = "grid";
-    private const string FrontCameraView = "front";
-    private const string RearCameraView = "rear";
-    private const string LeftCameraView = "left";
-    private const string RightCameraView = "right";
-
-    private readonly List<CamClip> _allClips = [];
-    private readonly FlyleafRuntime _flyleafRuntime = new();
-    private readonly UpdateService _updateService = new();
-    private VideoPlayerController _playerController;
-    private bool _isSeeking;
-    private bool _isInitialized;
+    private readonly MainWindowViewModel _viewModel;
     private bool _isClosing;
     private bool _isReadyToClose;
 
     public MainWindow()
     {
         InitializeComponent();
-        DataContext = this;
+
+        _viewModel = new MainWindowViewModel(
+            () => VideoPlayerController.Create(FrontFlyleafHost, BackFlyleafHost, LeftFlyleafHost, RightFlyleafHost));
+        _viewModel.SearchBoxFocusRequested += OnSearchBoxFocusRequested;
+        _viewModel.PropertyChanged += ViewModelOnPropertyChanged;
+
+        DataContext = _viewModel;
         UpdateCameraHostLayout();
     }
 
-    public bool ShowMainContent => !ShowAboutPage;
-
-    public Version CurrentVersion => Assembly.GetExecutingAssembly().GetName().Version ?? new Version(0, 0, 0);
-
-    public string FileVersion => FormatVersion(CurrentVersion);
-
-    public string RuntimeDescription => $"{RuntimeInformation.FrameworkDescription} ({RuntimeInformation.ProcessArchitecture})";
-
-    public string OsDescription => RuntimeInformation.OSDescription;
-
-    public string ExecutablePath => Environment.ProcessPath;
-
-    public bool HasUpdateBadge => IsUpdateAvailable;
-
-    public string LatestVersionText => LatestRelease is null ? "Unknown" : FormatVersion(LatestRelease.Version);
-
-    public string LatestReleaseUrl => LatestRelease?.ReleaseUrl ?? UpdateService.ReleasesPageUrl;
-
-    public string ReleasesPageUrl => UpdateService.ReleasesPageUrl;
-
-    public string UpdateStatusTitle => IsUpdateAvailable
-        ? "Update available"
-        : "You're up to date";
-
-    public string UpdateStatusDetails => IsUpdateAvailable
-        ? $"Version {LatestVersionText} is available."
-        : "No newer release was found.";
-
-    public IReadOnlyList<double> PlaybackSpeedOptions { get; } =
-    [
-        0.25,
-        0.5,
-        0.75,
-        1.0,
-        1.25,
-        1.5,
-        2.0,
-        3.0,
-        4.0,
-    ];
-
-    public IReadOnlyList<CamClip> FilteredClips => _allClips
-        .Where(c => string.IsNullOrWhiteSpace(FilterText) ||
-                    c.Name.Contains(FilterText, StringComparison.CurrentCultureIgnoreCase) ||
-                    c.FullPath.Contains(FilterText, StringComparison.CurrentCultureIgnoreCase))
-        .OrderByDescending(c => c.Timestamp)
-        .ThenBy(c => c.Name)
-        .ToList();
-
-    public string PositionText
-    {
-        get
-        {
-            var duration = _playerController?.Duration ?? TimeSpan.Zero;
-            var position = TimeSpan.FromSeconds(SeekPosition * duration.TotalSeconds);
-            return FormatTimeSpan(position);
-        }
-    }
-
-    public string DurationText => FormatTimeSpan(_playerController?.Duration ?? TimeSpan.Zero);
-
-    public bool CanSeek => _playerController?.IsMediaOpen == true && !IsLoading && _playerController.Duration > TimeSpan.Zero;
-
-    public bool CanPlayPause => (SelectedClip is not null || IsPlaying) && !IsLoading;
-
-    public bool CanStop => IsPlaying || IsLoading;
-
-    public bool CanGoNext => _playerController?.CanGoNext == true;
-
-    public bool CanGoPrevious => _playerController?.CanGoPrevious == true;
-
-    public string PlayPauseIcon => IsPlaying ? "\u23F8" : "\u25B6";
-
-    public bool ShowStatusOverlay => IsLoading || ShowErrorOverlay || HasNoClipSelected;
-
-    public bool ShowVideoHosts => !ShowStatusOverlay;
-
-    public bool HasError => ShowErrorOverlay;
-
-    public bool HasNoClipSelected => SelectedClip is null && !IsLoading && !ShowErrorOverlay;
-
-    public bool IsIndeterminateProgress => IsLoading && !IsRendering;
-
-    public bool IsGridViewSelected => SelectedCameraView == GridCameraView;
-
-    public bool IsSingleCameraViewSelected => !IsGridViewSelected;
-
-    public bool IsFrontViewSelected => SelectedCameraView == FrontCameraView;
-
-    public bool IsRearViewSelected => SelectedCameraView == RearCameraView;
-
-    public bool IsLeftViewSelected => SelectedCameraView == LeftCameraView;
-
-    public bool IsRightViewSelected => SelectedCameraView == RightCameraView;
-
-    public string ActiveCameraLabel => SelectedCameraView switch
-    {
-        GridCameraView => "Grid",
-        RearCameraView => "Rear",
-        LeftCameraView => "Left",
-        RightCameraView => "Right",
-        _ => "Front",
-    };
-
-    public string LoadingStatusText => IsRendering
-        ? $"Rendering... {RenderProgressPercent}%"
-        : "Loading...";
-
-    public int RenderProgressPercent => (int)(RenderProgress * 100);
-
-    [ObservableProperty]
-    [NotifyPropertyChangedFor(nameof(FilteredClips))]
-    private string _filterText = string.Empty;
-
-    [ObservableProperty]
-    [NotifyPropertyChangedFor(nameof(HasNoClipSelected))]
-    [NotifyPropertyChangedFor(nameof(ShowStatusOverlay))]
-    [NotifyPropertyChangedFor(nameof(ShowVideoHosts))]
-    [NotifyPropertyChangedFor(nameof(CanPlayPause))]
-    private CamClip _selectedClip;
-
-    [ObservableProperty]
-    private string _errorTitle;
-
-    [ObservableProperty]
-    private string _errorDetails;
-
-    [ObservableProperty]
-    [NotifyPropertyChangedFor(nameof(ShowStatusOverlay))]
-    [NotifyPropertyChangedFor(nameof(ShowVideoHosts))]
-    [NotifyPropertyChangedFor(nameof(HasNoClipSelected))]
-    [NotifyPropertyChangedFor(nameof(HasError))]
-    private bool _showErrorOverlay;
-
-    [ObservableProperty]
-    private bool _canDismissError = true;
-
-    [ObservableProperty]
-    private bool _showFFmpegDownloadButton;
-
-    [ObservableProperty]
-    [NotifyPropertyChangedFor(nameof(ShowMainContent))]
-    private bool _showAboutPage;
-
-    [ObservableProperty]
-    [NotifyPropertyChangedFor(nameof(CanPlayPause))]
-    [NotifyPropertyChangedFor(nameof(CanStop))]
-    [NotifyPropertyChangedFor(nameof(LoadingStatusText))]
-    [NotifyPropertyChangedFor(nameof(IsIndeterminateProgress))]
-    [NotifyPropertyChangedFor(nameof(ShowStatusOverlay))]
-    [NotifyPropertyChangedFor(nameof(ShowVideoHosts))]
-    [NotifyPropertyChangedFor(nameof(HasNoClipSelected))]
-    [NotifyPropertyChangedFor(nameof(CanSeek))]
-    private bool _isLoading;
-
-    [ObservableProperty]
-    [NotifyPropertyChangedFor(nameof(LoadingStatusText))]
-    [NotifyPropertyChangedFor(nameof(IsIndeterminateProgress))]
-    private bool _isRendering;
-
-    [ObservableProperty]
-    [NotifyPropertyChangedFor(nameof(RenderProgressPercent))]
-    [NotifyPropertyChangedFor(nameof(LoadingStatusText))]
-    private double _renderProgress;
-
-    [ObservableProperty]
-    [NotifyPropertyChangedFor(nameof(PositionText))]
-    private double _seekPosition;
-
-    [ObservableProperty]
-    [NotifyPropertyChangedFor(nameof(PlayPauseIcon))]
-    [NotifyPropertyChangedFor(nameof(CanPlayPause))]
-    [NotifyPropertyChangedFor(nameof(CanStop))]
-    private bool _isPlaying;
-
-    [ObservableProperty]
-    private double _selectedPlaybackSpeed = 1.0;
-
-    [ObservableProperty]
-    [NotifyPropertyChangedFor(nameof(IsGridViewSelected))]
-    [NotifyPropertyChangedFor(nameof(IsSingleCameraViewSelected))]
-    [NotifyPropertyChangedFor(nameof(IsFrontViewSelected))]
-    [NotifyPropertyChangedFor(nameof(IsRearViewSelected))]
-    [NotifyPropertyChangedFor(nameof(IsLeftViewSelected))]
-    [NotifyPropertyChangedFor(nameof(IsRightViewSelected))]
-    [NotifyPropertyChangedFor(nameof(ActiveCameraLabel))]
-    private string _selectedCameraView = FrontCameraView;
-
-    [ObservableProperty]
-    [NotifyPropertyChangedFor(nameof(HasUpdateBadge))]
-    [NotifyPropertyChangedFor(nameof(UpdateStatusTitle))]
-    [NotifyPropertyChangedFor(nameof(UpdateStatusDetails))]
-    private bool _isUpdateAvailable;
-
-    [ObservableProperty]
-    [NotifyPropertyChangedFor(nameof(LatestVersionText))]
-    [NotifyPropertyChangedFor(nameof(LatestReleaseUrl))]
-    [NotifyPropertyChangedFor(nameof(UpdateStatusDetails))]
-    private UpdateRelease _latestRelease;
-
     private async void Window_ContentRendered(object sender, EventArgs e)
     {
-        await InitializeAsync();
+        await _viewModel.InitializeAsync();
     }
 
     private void Window_Closing(object sender, CancelEventArgs e)
@@ -274,7 +63,7 @@ public partial class MainWindow : Window
     {
         try
         {
-            Shutdown();
+            _viewModel.Shutdown();
         }
         catch (Exception ex)
         {
@@ -289,7 +78,7 @@ public partial class MainWindow : Window
 
     private async void Window_KeyDown(object sender, KeyEventArgs e)
     {
-        if (await HandleKeyDownAsync(e.Key, Keyboard.Modifiers))
+        if (await _viewModel.HandleKeyDownAsync(e.Key, Keyboard.Modifiers))
         {
             e.Handled = true;
         }
@@ -297,496 +86,31 @@ public partial class MainWindow : Window
 
     private void SeekSlider_PreviewMouseDown(object sender, MouseButtonEventArgs e)
     {
-        BeginSeek();
+        _viewModel.BeginSeek();
     }
 
     private async void SeekSlider_PreviewMouseUp(object sender, MouseButtonEventArgs e)
     {
-        await EndSeekAsync();
+        await _viewModel.EndSeekAsync();
     }
 
     private void SeekSlider_ValueChanged(object sender, RoutedPropertyChangedEventArgs<double> e)
     {
-        if (_isSeeking)
+        _viewModel.OnSeekSliderValueChanged();
+    }
+
+    private void ViewModelOnPropertyChanged(object sender, PropertyChangedEventArgs e)
+    {
+        if (e.PropertyName == nameof(MainWindowViewModel.SelectedCameraView))
         {
-            OnPropertyChanged(nameof(PositionText));
+            UpdateCameraHostLayout();
         }
     }
 
-    private async Task InitializeAsync()
+    private void OnSearchBoxFocusRequested(object sender, EventArgs e)
     {
-        if (_isInitialized)
-            return;
-
-        _isInitialized = true;
-        Log.Debug("Initializing main window");
-
-#if DEBUG
-        Log.Debug("Skipping update check in debug build");
-#else
-        _ = UpdateLatestReleaseAsync();
-#endif
-
-        if (_flyleafRuntime.TryStart())
-        {
-            InitializePlayer();
-            LoadClips(CamStorage.FindCommonRoots());
-        }
-        else
-        {
-            ShowFFmpegMissingError();
-        }
-    }
-
-    private void Shutdown()
-    {
-        var controller = _playerController;
-        _playerController = null;
-
-        if (controller is null)
-            return;
-
-        controller.PropertyChanged -= PlayerControllerOnPropertyChanged;
-        controller.Dispose();
-    }
-
-    private void InitializePlayer()
-    {
-        if (_playerController is not null)
-            return;
-
-        _playerController = VideoPlayerController.Create(FrontFlyleafHost, BackFlyleafHost, LeftFlyleafHost, RightFlyleafHost);
-        _playerController.PropertyChanged += PlayerControllerOnPropertyChanged;
-        _playerController.PlaybackSpeed = SelectedPlaybackSpeed;
-    }
-
-    private void LoadClips(IEnumerable<string> roots)
-    {
-        ClearError();
-        _allClips.Clear();
-        SelectedClip = null;
-
-        var rootList = roots?.Where(root => !string.IsNullOrWhiteSpace(root)).ToList() ?? [];
-        if (rootList.Count == 0)
-        {
-            Log.Information("No dashcam roots found");
-            ShowError(
-                "No Dashcam Folders Found",
-                "Click 'Select Folder' to choose a folder containing Tesla dashcam footage (TeslaCam folder).",
-                canDismiss: true);
-            RefreshClipState();
-            return;
-        }
-
-        Log.Information("Loading dashcam clips. RootCount={RootCount}; Roots={Roots}", rootList.Count, rootList);
-        var totalStopwatch = Stopwatch.StartNew();
-        var failedRoots = 0;
-
-        foreach (var root in rootList)
-        {
-            var rootStopwatch = Stopwatch.StartNew();
-            Log.Debug("Scanning dashcam root. Root={Root}", root);
-
-            try
-            {
-                var storage = CamStorage.Map(root);
-                _allClips.AddRange(storage.Clips);
-                Log.Information(
-                    "Scanned dashcam root. Root={Root}; ClipCount={ClipCount}; ElapsedMs={ElapsedMs}",
-                    root,
-                    storage.Clips.Count,
-                    rootStopwatch.ElapsedMilliseconds);
-            }
-            catch (UnauthorizedAccessException ex)
-            {
-                failedRoots++;
-                Log.Error(ex, "Access denied while loading dashcam root. Root={Root}", root);
-                ShowError("Access Denied", $"Cannot access folder: {root}\n\nCheck that you have permission to read this location.");
-            }
-            catch (Exception ex)
-            {
-                failedRoots++;
-                Log.Error(ex, "Failed to load dashcam root. Root={Root}", root);
-                ShowError("Error Loading Clips", $"Failed to load clips from:\n{root}\n\nError: {ex.Message}");
-            }
-        }
-
-        _playerController?.LoadClips(_allClips);
-        RefreshClipState();
-        Log.Information(
-            "Finished loading dashcam clips. ClipCount={ClipCount}; RootCount={RootCount}; FailedRootCount={FailedRootCount}; ElapsedMs={ElapsedMs}",
-            _allClips.Count,
-            rootList.Count,
-            failedRoots,
-            totalStopwatch.ElapsedMilliseconds);
-    }
-
-    private void RefreshClipState()
-    {
-        OnPropertyChanged(nameof(FilteredClips));
-        OnPropertyChanged(nameof(HasNoClipSelected));
-        OnPropertyChanged(nameof(ShowStatusOverlay));
-        OnPropertyChanged(nameof(ShowVideoHosts));
-        OnPropertyChanged(nameof(CanPlayPause));
-        OnPropertyChanged(nameof(CanGoNext));
-        OnPropertyChanged(nameof(CanGoPrevious));
-    }
-
-    [RelayCommand]
-    private async Task OpenFolderAsync()
-    {
-        Log.Debug("Opening folder picker");
-
-        var dialog = new OpenFolderDialog
-        {
-            Multiselect = true,
-            Title = "Select a folder containing Tesla dashcam footage (TeslaCam folder)",
-        };
-
-        if (dialog.ShowDialog() == true)
-        {
-            Log.Information(
-                "User selected dashcam folders. FolderCount={FolderCount}; Folders={Folders}",
-                dialog.FolderNames.Length,
-                dialog.FolderNames);
-
-            if (_playerController is not null)
-            {
-                await _playerController.StopAsync();
-            }
-
-            LoadClips(dialog.FolderNames);
-        }
-        else
-        {
-            Log.Debug("Folder picker canceled");
-        }
-    }
-
-    [RelayCommand]
-    private async Task PlayPauseAsync()
-    {
-        if (_playerController is not null)
-        {
-            await _playerController.TogglePlayPauseAsync();
-        }
-    }
-
-    [RelayCommand]
-    private async Task StopAsync()
-    {
-        if (_playerController is null)
-            return;
-
-        await _playerController.StopAsync();
-        SeekPosition = 0;
-    }
-
-    [RelayCommand]
-    private async Task PreviousAsync()
-    {
-        if (_playerController is null)
-            return;
-
-        await _playerController.PreviousAsync();
-        SelectedClip = _playerController.CurrentClip;
-    }
-
-    [RelayCommand]
-    private async Task NextAsync()
-    {
-        if (_playerController is null)
-            return;
-
-        await _playerController.NextAsync();
-        SelectedClip = _playerController.CurrentClip;
-    }
-
-    [RelayCommand]
-    private async Task DownloadFFmpegAsync()
-    {
-        IsLoading = true;
-        ClearError();
-
-        try
-        {
-            Log.Debug("Starting FFmpeg download workflow");
-            await PackageManager.DownloadAndExtractFFmpeg();
-            if (_flyleafRuntime.TryStart())
-            {
-                InitializePlayer();
-                LoadClips(CamStorage.FindCommonRoots());
-            }
-            else
-            {
-                ShowFFmpegMissingError();
-            }
-        }
-        catch (Exception ex)
-        {
-            Log.Error(ex, "Failed to download FFmpeg");
-            ShowError("Download Failed", $"Failed to download FFmpeg: {ex.Message}");
-            ShowFFmpegDownloadButton = true;
-        }
-        finally
-        {
-            IsLoading = false;
-        }
-    }
-
-    private async Task PlaySelectedClipAsync()
-    {
-        if (SelectedClip is null || _playerController is null)
-            return;
-
-        ClearError();
-        IsLoading = true;
-        await Dispatcher.Yield(DispatcherPriority.Background);
-
-        try
-        {
-            await _playerController.GoToClipAsync(SelectedClip);
-        }
-        catch (Exception ex)
-        {
-            IsLoading = false;
-            Log.Error(
-                ex,
-                "Failed to play selected clip. ClipName={ClipName}; ClipPath={ClipPath}",
-                SelectedClip.Name,
-                SelectedClip.FullPath);
-            ShowError("Playback Failed", $"Could not play clip: {SelectedClip.Name}\n\nError: {ex.Message}");
-        }
-    }
-
-    private async Task<bool> HandleKeyDownAsync(Key key, ModifierKeys modifiers)
-    {
-        if ((key == Key.F && modifiers == ModifierKeys.Control) ||
-            (key == Key.F3 && modifiers == ModifierKeys.None))
-        {
-            FocusSearchBox();
-            return true;
-        }
-
-        if (_playerController is null)
-        {
-            return false;
-        }
-
-        switch (key)
-        {
-            case Key.Space:
-                await _playerController.TogglePlayPauseAsync();
-                return true;
-
-            case Key.Left:
-                if (modifiers == ModifierKeys.Control && CanGoPrevious)
-                {
-                    await _playerController.PreviousAsync();
-                    SelectedClip = _playerController.CurrentClip;
-                }
-                else if (CanSeek)
-                {
-                    var position = _playerController.Position - TimeSpan.FromSeconds(5);
-                    await _playerController.SeekAsync(position < TimeSpan.Zero ? TimeSpan.Zero : position);
-                }
-
-                return true;
-
-            case Key.Right:
-                if (modifiers == ModifierKeys.Control && CanGoNext)
-                {
-                    await _playerController.NextAsync();
-                    SelectedClip = _playerController.CurrentClip;
-                }
-                else if (CanSeek)
-                {
-                    var duration = _playerController.Duration;
-                    var position = _playerController.Position + TimeSpan.FromSeconds(5);
-                    await _playerController.SeekAsync(position > duration ? duration : position);
-                }
-
-                return true;
-        }
-
-        return false;
-    }
-
-    private void FocusSearchBox()
-    {
-        ShowAboutPage = false;
         SearchBox.Focus();
         SearchBox.SelectAll();
-    }
-
-    private void BeginSeek()
-    {
-        if (CanSeek)
-        {
-            _isSeeking = true;
-        }
-    }
-
-    private async Task EndSeekAsync()
-    {
-        if (_playerController is null || !CanSeek)
-        {
-            _isSeeking = false;
-            return;
-        }
-
-        await SeekToCurrentPositionAsync();
-        _isSeeking = false;
-    }
-
-    private async Task SeekToCurrentPositionAsync()
-    {
-        if (_playerController is null)
-            return;
-
-        var duration = _playerController.Duration;
-        if (duration.TotalSeconds > 0)
-        {
-            var targetPosition = TimeSpan.FromSeconds(SeekPosition * duration.TotalSeconds);
-            await _playerController.SeekAsync(targetPosition);
-        }
-    }
-
-    private void UpdateSeekPositionFromController()
-    {
-        if (_playerController is null || _isSeeking)
-            return;
-
-        var duration = _playerController.Duration;
-        SeekPosition = duration.TotalSeconds > 0
-            ? Math.Clamp(_playerController.Position.TotalSeconds / duration.TotalSeconds, 0, 1)
-            : 0;
-    }
-
-    private void PlayerControllerOnPropertyChanged(object sender, PropertyChangedEventArgs e)
-    {
-        if (string.IsNullOrWhiteSpace(e.PropertyName))
-            return;
-
-        RunOnUiThread(() => HandlePlayerControllerPropertyChanged(e.PropertyName));
-    }
-
-    private void HandlePlayerControllerPropertyChanged(string propertyName)
-    {
-        if (_playerController is null)
-            return;
-
-        switch (propertyName)
-        {
-            case nameof(VideoPlayerController.IsLoading):
-                IsLoading = _playerController.IsLoading;
-                break;
-
-            case nameof(VideoPlayerController.IsPlaying):
-                IsPlaying = _playerController.IsPlaying;
-                break;
-
-            case nameof(VideoPlayerController.Duration):
-                UpdateSeekPositionFromController();
-                OnPropertyChanged(nameof(DurationText));
-                OnPropertyChanged(nameof(PositionText));
-                OnPropertyChanged(nameof(CanSeek));
-                break;
-
-            case nameof(VideoPlayerController.Position):
-                UpdateSeekPositionFromController();
-                break;
-
-            case nameof(VideoPlayerController.ErrorMessage):
-                if (_playerController.ErrorMessage is not null)
-                {
-                    ShowError("Playback Error", _playerController.ErrorMessage);
-                }
-
-                break;
-
-            case nameof(VideoPlayerController.CurrentClip):
-                SelectedClip = _playerController.CurrentClip;
-                RefreshClipState();
-                break;
-
-            case nameof(VideoPlayerController.IsMediaOpen):
-                OnPropertyChanged(nameof(CanSeek));
-                break;
-        }
-    }
-
-    internal static string FormatTimeSpan(TimeSpan ts)
-    {
-        return ts.TotalHours >= 1
-            ? ts.ToString(@"h\:mm\:ss")
-            : ts.ToString(@"m\:ss");
-    }
-
-    private static string FormatVersion(Version version)
-    {
-        if (version is null)
-        {
-            return "Unknown";
-        }
-
-        if (version.Revision >= 0)
-        {
-            return version.ToString(4);
-        }
-
-        return version.Build >= 0
-            ? version.ToString(3)
-            : version.ToString(2);
-    }
-
-    private void ShowError(string title, string details, bool canDismiss = true)
-    {
-        ErrorTitle = title;
-        ErrorDetails = details;
-        CanDismissError = canDismiss;
-        ShowErrorOverlay = true;
-    }
-
-    private void ClearError()
-    {
-        ShowErrorOverlay = false;
-        ShowFFmpegDownloadButton = false;
-        CanDismissError = true;
-        ErrorTitle = null;
-        ErrorDetails = null;
-    }
-
-    [RelayCommand]
-    private void DismissError()
-    {
-        ClearError();
-    }
-
-    private void ShowFFmpegMissingError()
-    {
-        Log.Debug("Showing FFmpeg missing prompt");
-        ShowFFmpegDownloadButton = true;
-        ShowError("FFmpeg Required", "FFmpeg is required to play clips. This will download about 80MB.", canDismiss: false);
-    }
-
-    [RelayCommand]
-    private void ToggleAbout()
-    {
-        ShowAboutPage = !ShowAboutPage;
-    }
-
-    [RelayCommand]
-    private void SelectCameraView(string cameraView)
-    {
-        SelectedCameraView = cameraView switch
-        {
-            GridCameraView => GridCameraView,
-            RearCameraView => RearCameraView,
-            LeftCameraView => LeftCameraView,
-            RightCameraView => RightCameraView,
-            _ => FrontCameraView,
-        };
     }
 
     private void UpdateCameraHostLayout()
@@ -796,30 +120,30 @@ public partial class MainWindow : Window
 
         ClearCameraHostSlots();
 
-        switch (SelectedCameraView)
+        switch (_viewModel.SelectedCameraView)
         {
-            case GridCameraView:
+            case MainWindowViewModel.GridCameraView:
                 MoveHostToSlot(FrontFlyleafHost, GridFrontHostSlot);
                 MoveHostToSlot(BackFlyleafHost, GridRearHostSlot);
                 MoveHostToSlot(LeftFlyleafHost, GridLeftHostSlot);
                 MoveHostToSlot(RightFlyleafHost, GridRightHostSlot);
                 break;
 
-            case RearCameraView:
+            case MainWindowViewModel.RearCameraView:
                 MoveHostToSlot(BackFlyleafHost, PrimaryCameraHostSlot);
                 MoveHostToSlot(FrontFlyleafHost, FrontTileHostSlot);
                 MoveHostToSlot(LeftFlyleafHost, LeftTileHostSlot);
                 MoveHostToSlot(RightFlyleafHost, RightTileHostSlot);
                 break;
 
-            case LeftCameraView:
+            case MainWindowViewModel.LeftCameraView:
                 MoveHostToSlot(LeftFlyleafHost, PrimaryCameraHostSlot);
                 MoveHostToSlot(FrontFlyleafHost, FrontTileHostSlot);
                 MoveHostToSlot(BackFlyleafHost, RearTileHostSlot);
                 MoveHostToSlot(RightFlyleafHost, RightTileHostSlot);
                 break;
 
-            case RightCameraView:
+            case MainWindowViewModel.RightCameraView:
                 MoveHostToSlot(RightFlyleafHost, PrimaryCameraHostSlot);
                 MoveHostToSlot(FrontFlyleafHost, FrontTileHostSlot);
                 MoveHostToSlot(BackFlyleafHost, RearTileHostSlot);
@@ -886,62 +210,6 @@ public partial class MainWindow : Window
         }
     }
 
-    private async Task UpdateLatestReleaseAsync()
-    {
-        var result = await _updateService.CheckForUpdateAsync(CurrentVersion);
-        LatestRelease = result.LatestRelease;
-        IsUpdateAvailable = result.IsUpdateAvailable;
-
-        Log.Information(
-            "Checked for updates. CurrentVersion={CurrentVersion}; LatestVersion={LatestVersion}; IsUpdateAvailable={IsUpdateAvailable}",
-            FormatVersion(CurrentVersion),
-            LatestVersionText,
-            IsUpdateAvailable);
-    }
-
-    private static bool CanUseClip(CamClip clip)
-    {
-        return clip is not null;
-    }
-
-    [RelayCommand(CanExecute = nameof(CanUseClip))]
-    private void OpenClipFolder(CamClip clip)
-    {
-        if (clip is null)
-        {
-            return;
-        }
-
-        if (!Directory.Exists(clip.FullPath))
-        {
-            ShowError("Clip Folder Not Found", $"Could not find folder:\n{clip.FullPath}");
-            return;
-        }
-
-        Process.Start(new ProcessStartInfo(clip.FullPath)
-        {
-            UseShellExecute = true,
-        });
-    }
-
-    [RelayCommand(CanExecute = nameof(CanUseClip))]
-    private void CopyClipPath(CamClip clip)
-    {
-        if (clip is not null)
-        {
-            Clipboard.SetText(clip.FullPath);
-        }
-    }
-
-    [RelayCommand(CanExecute = nameof(CanUseClip))]
-    private void CopyClipName(CamClip clip)
-    {
-        if (clip is not null)
-        {
-            Clipboard.SetText(clip.Name);
-        }
-    }
-
     private void Hyperlink_RequestNavigate(object sender, RequestNavigateEventArgs e)
     {
         if (e.Uri is null)
@@ -955,42 +223,5 @@ public partial class MainWindow : Window
         });
 
         e.Handled = true;
-    }
-
-    private void RunOnUiThread(Action action)
-    {
-        if (Dispatcher.CheckAccess())
-        {
-            action();
-            return;
-        }
-
-        Dispatcher.Invoke(action);
-    }
-
-    partial void OnSelectedClipChanged(CamClip value)
-    {
-        if (value is not null)
-        {
-            Log.Debug(
-                "Selected clip changed. ClipName={ClipName}; ClipPath={ClipPath}",
-                value.Name,
-                value.FullPath);
-            _ = PlaySelectedClipAsync();
-        }
-    }
-
-    partial void OnSelectedPlaybackSpeedChanged(double value)
-    {
-        if (_playerController is not null)
-        {
-            Log.Information("Playback speed changed. Speed={PlaybackSpeed}", value);
-            _playerController.PlaybackSpeed = value;
-        }
-    }
-
-    partial void OnSelectedCameraViewChanged(string value)
-    {
-        UpdateCameraHostLayout();
     }
 }
